refactor: **Filter를 **Query로 이름 수정
- 인터페이스마다 수행해야하는 쿼리가 정의되어있음으로 보다 표시성이 높은 이름으로 수정
