+++ persistence/src/main/java/com/comssa/persistence/question/repository/querydsl/query/MajorQuestionSearchQuery.java
@@ -1,14 +1,15 @@
-package com.comssa.persistence.question.repository.querydsl;
+package com.comssa.persistence.question.repository.querydsl.query;
 
 import com.comssa.persistence.question.domain.common.Question;
 import com.comssa.persistence.question.domain.common.QuestionCategory;
 import com.comssa.persistence.question.domain.common.QuestionLevel;
+import com.comssa.persistence.question.repository.querydsl.query.booleanbuilder.LevelsAndCategoryBooleanBuilder;
 import com.querydsl.jpa.impl.JPAQuery;
 
 import java.util.List;
 
-public interface MajorQuestionSearchFilter<T extends Question>
-	extends LevelsAndCategoryBooleanBuilder, CommonQuestionFilter<T> {
+public interface MajorQuestionSearchQuery<T extends Question>
+	extends LevelsAndCategoryBooleanBuilder, QuestionQuerySupport<T> {
 
 	List<T> findAllCategoriesAndLevelsAndIfApproved(
 		List<QuestionCategory> questionCategories,
@@ -16,6 +17,12 @@ public interface MajorQuestionSearchFilter<T extends Question>
 		boolean approved
 	);
 
+	List<T> findAllOrderByIfApprovedAsc();
+
+	default JPAQuery<T> selectOrderByIfApprovedAsc() {
+		return getQuestion()
+			.orderBy(getQuestionQClass().ifApproved.asc());
+	}
 
 	default JPAQuery<T> selectWhereCategoriesAndLevelsAndIfApproved(
 		List<QuestionCategory> questionCategories,
